renamed the file to match the KML exporter class it mimics #243
